// angular-client/src/app/server/model/responseAccount.ts
/**
 * Accounting
 * This is an accounting server. 
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ResponseAccount { 
    id?: number;
    name?: string;
    parentId?: number;
    active?: boolean;
}